feat: handle jsx rendering order smartly to improve UX & remove hacks 🔥🔥
instead of reversing the items we want to render and rendering after the target,
we capture the target's next node and always render before it.

This matters because of the rendering order -- the earlier you render,
the earlier a request is made (if required) for data retrieval etc.,
and the quicker the element finishes loading.

An issue became very obvious after implementing the active loading indicator --
since we reversed the custom label picker elements, the visually-lower ones
would resolve their data faster, and the visually-higher ones would be slower.

This proved to be very unsatisfying to experience from the UX POV,
and thus this change came along 🚀